:axe: helper per file (almost)
