+++ src/hepers/pipe.ts
@@ -1,0 +1,4 @@
+const pipe = (value: any, ...fns: Function[]) =>
+    fns.reduce((acc, fn) => fn(acc), value);
+
+export { pipe }
